requirements.txt
pydicom
matplotlib

